- Deleted the old classes that are no more used - Restructured the packages
