+++ src/me/purplefishh/dipcraft/superbet/command/HelpCommand.java
@@ -1,5 +1,6 @@
 package me.purplefishh.dipcraft.superbet.command;
 
+import me.purplefishh.dipcraft.superbet.command.system.ICommand;
 import me.purplefishh.dipcraft.superbet.configCollections.MessagesCollection;
 import org.bukkit.command.CommandSender;
