--- src/me/purplefishh/dipcraft/superbet/resorce/ColorUtils.java
@@ -1,7 +1,0 @@
-package me.purplefishh.dipcraft.superbet.resorce;
-
-public class ColorUtils {
-
-
-
-}
